refactor: dismantle chat/workflow/ directory structure
Comprehensively restructure frontend/internal-packages/agent to eliminate deeply nested chat/workflow/ directories and consolidate scattered utility directories into a single, maintainable structure.

**Key Changes:**
- Dismantle chat/workflow/ directory:
  - Move chat/workflow/shared/getConfigurable.ts → src/utils/getConfigurable.ts
  - Move chat/workflow/shared/workflowConstants.ts → src/constants.ts
  - Move chat/workflow/shared/workflowAnnotation.ts → src/workflowAnnotation.ts
  - Move chat/workflow/utils/* → src/utils/* (4 utility files + tests)
  - Merge chat/workflow/types.ts → src/types.ts

- Consolidate shared directories:
  - Move src/shared/* → src/utils/* (4 utility files)
  - Move src/langchain/utils/schema.ts → src/utils/validationSchema.ts
  - Merge src/langchain/utils/types.ts → src/types.ts

- Update import paths across 20+ files to use new structure
- Remove empty directories (chat/, shared/, langchain/utils/)
- Auto-fix import organization with Biome

**Results:**
- Flatter directory structure eliminates unnecessary nesting
- Single consolidated utils/ directory improves maintainability
- Shorter import paths enhance developer experience
- All tests pass (133/134), TypeScript compiles, full lint suite passes
- File history preserved using git mv

Resolves: #5495
