+++ frontend/internal-packages/agent/src/createGraph.ts
@@ -1,13 +1,13 @@
 import type { RunnableConfig } from '@langchain/core/runnables'
 import { END, START, StateGraph } from '@langchain/langgraph'
 import type { BaseCheckpointSaver } from '@langchain/langgraph-checkpoint'
-import { workflowAnnotation } from './chat/workflow/shared/workflowAnnotation'
-import type { WorkflowState } from './chat/workflow/types'
 import { createDbAgentGraph } from './db-agent/createDbAgentGraph'
 import { createLeadAgentGraph } from './lead-agent/createLeadAgentGraph'
 import { createPmAgentGraph } from './pm-agent/createPmAgentGraph'
 import { createQaAgentGraph } from './qa-agent/createQaAgentGraph'
+import type { WorkflowState } from './types'
 import { validateInitialSchemaNode } from './workflow/nodes/validateInitialSchemaNode'
+import { workflowAnnotation } from './workflowAnnotation'
 
 /**
  * Create and configure the LangGraph workflow
