+++ frontend/internal-packages/agent/src/db-agent/invokeDesignAgent.ts
@@ -11,8 +11,8 @@ import { ResultAsync } from 'neverthrow'
 import { v4 as uuidv4 } from 'uuid'
 import * as v from 'valibot'
 import { SSE_EVENTS } from '../client'
-import { reasoningSchema } from '../langchain/utils/schema'
-import type { Reasoning } from '../langchain/utils/types'
+import type { Reasoning } from '../types'
+import { reasoningSchema } from '../utils/validationSchema'
 import type { ToolConfigurable } from './getToolConfigurable'
 import { type DesignAgentPromptVariables, designAgentPrompt } from './prompt'
 import { schemaDesignTool } from './tools/schemaDesignTool'
