+++ frontend/internal-packages/agent/src/db-agent/nodes/invokeSchemaDesignToolNode.ts
@@ -4,8 +4,8 @@ import type { RunnableConfig } from '@langchain/core/runnables'
 import { ToolNode } from '@langchain/langgraph/prebuilt'
 import type { Schema } from '@liam-hq/schema'
 import type { ResultAsync } from 'neverthrow'
-import { getConfigurable } from '../../chat/workflow/shared/getConfigurable'
 import type { Repositories } from '../../repositories'
+import { getConfigurable } from '../../utils/getConfigurable'
 import type { DbAgentState } from '../shared/dbAgentAnnotation'
 import { schemaDesignTool } from '../tools/schemaDesignTool'
