+++ frontend/internal-packages/agent/src/db-agent/routing/routeAfterDesignSchema.test.ts
@@ -1,7 +1,7 @@
 import { AIMessage, HumanMessage } from '@langchain/core/messages'
 import { END } from '@langchain/langgraph'
 import { describe, expect, it } from 'vitest'
-import type { WorkflowState } from '../../chat/workflow/types'
+import type { WorkflowState } from '../../types'
 import { routeAfterDesignSchema } from './routeAfterDesignSchema'
 
 const workflowState = (messages: WorkflowState['messages']): WorkflowState => ({
