+++ frontend/internal-packages/agent/src/deepModeling.ts
@@ -1,11 +1,14 @@
-import { DEFAULT_RECURSION_LIMIT } from './chat/workflow/shared/workflowConstants'
-import type { WorkflowConfigurable } from './chat/workflow/types'
+import { DEFAULT_RECURSION_LIMIT } from './constants'
 import { createGraph } from './createGraph'
+import type {
+  AgentWorkflowParams,
+  AgentWorkflowResult,
+  WorkflowConfigurable,
+} from './types'
 import {
   executeWorkflowWithTracking,
   setupWorkflowState,
-} from './shared/workflowSetup'
-import type { AgentWorkflowParams, AgentWorkflowResult } from './types'
+} from './utils/workflowSetup'
 
 /**
  * Execute Deep Modeling workflow
