+++ frontend/internal-packages/agent/src/index.ts
@@ -5,6 +5,6 @@ export * from './deepModeling'
 export * from './deepModelingStream'
 export * from './getMessages'
 export * from './repositories'
-export * from './shared/workflowSetup'
 export * from './types'
 export * from './utils/nodeLogger'
+export * from './utils/workflowSetup'
