--- frontend/internal-packages/agent/src/langchain/utils/types.ts
@@ -1,4 +1,0 @@
-import type * as v from 'valibot'
-import type { reasoningSchema } from './schema'
-
-export type Reasoning = v.InferOutput<typeof reasoningSchema>
