+++ frontend/internal-packages/agent/src/lead-agent/classifyMessage/index.integration.test.ts
@@ -10,8 +10,8 @@ import {
   getTestConfig,
   outputStream,
 } from '../../../test-utils/workflowTestHelpers'
-import { workflowAnnotation } from '../../chat/workflow/shared/workflowAnnotation'
-import type { WorkflowState } from '../../chat/workflow/types'
+import type { WorkflowState } from '../../types'
+import { workflowAnnotation } from '../../workflowAnnotation'
 import { classifyMessage } from './index'
 
 const createTestStateAndConfig = async (
