+++ frontend/internal-packages/agent/src/lead-agent/summarizeWorkflow/index.integration.test.ts
@@ -5,8 +5,8 @@ import {
   getTestConfig,
   outputStream,
 } from '../../../test-utils/workflowTestHelpers'
-import { workflowAnnotation } from '../../chat/workflow/shared/workflowAnnotation'
-import type { WorkflowState } from '../../chat/workflow/types'
+import type { WorkflowState } from '../../types'
+import { workflowAnnotation } from '../../workflowAnnotation'
 import { summarizeWorkflow } from './index'
 
 describe('summarizeWorkflow Integration', () => {
