+++ frontend/internal-packages/agent/src/lead-agent/summarizeWorkflow/index.ts
@@ -8,8 +8,8 @@ import { END } from '@langchain/langgraph'
 import { ChatOpenAI } from '@langchain/openai'
 import { ResultAsync } from 'neverthrow'
 import { v4 as uuidv4 } from 'uuid'
-import type { WorkflowState } from '../../chat/workflow/types'
 import { SSE_EVENTS } from '../../client'
+import type { WorkflowState } from '../../types'
 
 const AGENT_NAME = 'lead' as const
