+++ frontend/internal-packages/agent/src/lead-agent/utils/workflowStatus.ts
@@ -1,4 +1,4 @@
-import type { WorkflowState } from '../../chat/workflow/types'
+import type { WorkflowState } from '../../types'
 
 export function isQACompleted(state: WorkflowState): boolean {
   return state.testcases.length > 0
