+++ frontend/internal-packages/agent/src/pm-agent/nodes/analyzeRequirementsNode.ts
@@ -1,7 +1,7 @@
 import type { RunnableConfig } from '@langchain/core/runnables'
-import { getConfigurable } from '../../chat/workflow/shared/getConfigurable'
-import { WorkflowTerminationError } from '../../shared/errorHandling'
 import { convertSchemaToText } from '../../utils/convertSchemaToText'
+import { WorkflowTerminationError } from '../../utils/errorHandling'
+import { getConfigurable } from '../../utils/getConfigurable'
 import { invokePmAnalysisAgent } from '../invokePmAnalysisAgent'
 import type { PmAgentState } from '../pmAgentAnnotations'
