+++ frontend/internal-packages/agent/src/pm-agent/tools/saveRequirementsToArtifactTool.ts
@@ -12,11 +12,11 @@ import { fromValibotSafeParse } from '@liam-hq/neverthrow'
 import { err, ok, type Result } from 'neverthrow'
 import { v4 as uuidv4 } from 'uuid'
 import * as v from 'valibot'
-import { getConfigurable } from '../../chat/workflow/shared/getConfigurable'
 import { SSE_EVENTS } from '../../client'
 import type { Repositories } from '../../repositories'
-import { WorkflowTerminationError } from '../../shared/errorHandling'
-import { toJsonSchema } from '../../shared/jsonSchema'
+import { WorkflowTerminationError } from '../../utils/errorHandling'
+import { getConfigurable } from '../../utils/getConfigurable'
+import { toJsonSchema } from '../../utils/jsonSchema'
 
 // Valibot schema for validating analyzedRequirements structure
 const analyzedRequirementsSchema = v.object({
