+++ frontend/internal-packages/agent/src/qa-agent/createQaAgentGraph.ts
@@ -1,6 +1,6 @@
 import { END, START, StateGraph } from '@langchain/langgraph'
 import type { BaseCheckpointSaver } from '@langchain/langgraph-checkpoint'
-import { RETRY_POLICY } from '../shared/errorHandling'
+import { RETRY_POLICY } from '../utils/errorHandling'
 import { generateTestcaseAndDmlNode } from './generateTestcaseAndDml'
 import { invokeSaveTestcasesAndDmlToolNode } from './nodes/invokeSaveTestcasesAndDmlToolNode'
 import { routeAfterGenerateTestcaseAndDml } from './routing/routeAfterGenerateTestcaseAndDml'
