+++ frontend/internal-packages/agent/src/qa-agent/generateTestcaseAndDml/index.ts
@@ -5,8 +5,8 @@ import {
 } from '@langchain/core/messages'
 import { ChatOpenAI } from '@langchain/openai'
 import { ResultAsync } from 'neverthrow'
-import { WorkflowTerminationError } from '../../shared/errorHandling'
 import { convertSchemaToText } from '../../utils/convertSchemaToText'
+import { WorkflowTerminationError } from '../../utils/errorHandling'
 import { removeReasoningFromMessages } from '../../utils/messageCleanup'
 import type { QaAgentState } from '../shared/qaAgentAnnotation'
 import { saveTestcasesAndDmlTool } from '../tools/saveTestcasesAndDmlTool'
