+++ frontend/internal-packages/agent/src/qa-agent/validateSchema/index.ts
@@ -4,10 +4,10 @@ import type { DmlOperation } from '@liam-hq/artifact'
 import { executeQuery } from '@liam-hq/pglite-server'
 import type { SqlResult } from '@liam-hq/pglite-server/src/types'
 import { ResultAsync } from 'neverthrow'
-import { getConfigurable } from '../../chat/workflow/shared/getConfigurable'
-import { generateDdlFromSchema } from '../../chat/workflow/utils/generateDdl'
-import { transformWorkflowStateToArtifact } from '../../chat/workflow/utils/transformWorkflowStateToArtifact'
-import { WorkflowTerminationError } from '../../shared/errorHandling'
+import { WorkflowTerminationError } from '../../utils/errorHandling'
+import { generateDdlFromSchema } from '../../utils/generateDdl'
+import { getConfigurable } from '../../utils/getConfigurable'
+import { transformWorkflowStateToArtifact } from '../../utils/transformWorkflowStateToArtifact'
 import type { QaAgentState } from '../shared/qaAgentAnnotation'
 import type { Testcase } from '../types'
 import { formatValidationErrors } from './formatValidationErrors'
