+++ frontend/internal-packages/agent/src/utils/formatWebSearchContent.test.ts
