+++ frontend/internal-packages/agent/src/utils/formatWebSearchContent.ts
