+++ frontend/internal-packages/agent/src/utils/generateDdl.ts
