+++ frontend/internal-packages/agent/src/utils/getConfigurable.ts
