+++ frontend/internal-packages/agent/src/utils/jsonSchema.ts
