+++ frontend/internal-packages/agent/src/utils/toolMessageUtils.test.ts
