+++ frontend/internal-packages/agent/src/utils/toolMessageUtils.ts
