+++ frontend/internal-packages/agent/src/utils/traceEnhancer.ts
