+++ frontend/internal-packages/agent/src/workflow/nodes/validateInitialSchemaNode.ts
@@ -1,5 +1,5 @@
 import type { RunnableConfig } from '@langchain/core/runnables'
-import type { WorkflowState } from '../../chat/workflow/types'
+import type { WorkflowState } from '../../types'
 
 /**
  * Validates initial schema and provides Instant Database initialization experience.
